+++ roles/files/lookup_plugins/manala_files_attributes.py
@@ -9,6 +9,19 @@ import re
 
 class LookupModule(LookupBase):
 
+    def _recursive_flatten(self, terms):
+        if not isinstance(terms, list):
+            raise AnsibleError('Expect a list but was a %s' % type(terms))
+        ret = []
+        for term in terms:
+            if isinstance(term, (list, tuple)):
+                ret.extend(
+                    self._recursive_flatten(term)
+                )
+            else:
+                ret.append(term)
+        return ret
+
     def _default(self, defaults, path):
         result = {}
         for default in defaults:
@@ -20,118 +33,53 @@ class LookupModule(LookupBase):
 
         results = []
 
-        attributes = self._flatten(terms[0])
+        attributes = self._recursive_flatten(terms[0])
         defaults = self._flatten(terms[1])
+
+        itemDefault = {}
 
         for attribute in attributes:
 
-            # Check index key
-            if 'path' not in attribute:
-                raise AnsibleError('Expect "path" key')
-
             items = []
 
-            # State - Link Directory
-            if 'state' in attribute and (attribute['state'] == 'link_directory'):
-                if 'src' not in attribute:
-                    raise AnsibleError('Expect "src" key')
-                item = {
-                    'path': attribute['path'],
-                    'src': attribute['src'],
-                    'link': {
-                        'parents': False,
-                        'force': False,
-                    },
-                    'directory': {
-                        'force': False,
-                    },
-                    'task': 'link_directory'
-                }
-                item['link'].update(self._default(defaults, attribute['path']))
-                item['link'].update(attribute)
-                item['directory'].update(self._default(defaults, attribute['src']))
-                item['directory'].update(attribute)
-                items.append(item)
-            # State - Link File
-            elif 'state' in attribute and (attribute['state'] == 'link_file'):
-                if 'src' not in attribute:
-                    raise AnsibleError('Expect "src" key')
-                item = {
-                    'path': attribute['path'],
-                    'src': attribute['src'],
-                    'link': {
-                        'parents': False,
-                        'force': False,
-                    },
-                    'file': {
-                        'parents': False,
-                        'force': False,
-                    },
-                    'task': 'link_file'
-                }
-                item['link'].update(self._default(defaults, attribute['path']))
-                item['link'].update(attribute)
-                item['file'].update(self._default(defaults, attribute['path']))
-                item['file'].update(attribute)
-                items.append(item)
-            else:
-                # Template
-                if 'template' in attribute:
-                    item = {
-                        'state': 'present',
-                        'parents': False
-                    }
-                    item.update(self._default(defaults, attribute['path']))
-                    item.update(attribute)
-                    item.update({
-                        'task': 'template'
-                    })
-                    items.append(item)
-                # Content
-                elif 'content' in attribute:
-                    item = {
-                        'state': 'present',
-                        'parents': False
-                    }
-                    item.update(self._default(defaults, attribute['path']))
-                    item.update(attribute)
-                    item.update({
-                        'task': 'content'
-                    })
-                    items.append(item)
-                # Copy
-                elif 'copy' in attribute:
-                    item = {
-                        'state': 'present'
-                    }
-                    item.update(self._default(defaults, attribute['path']))
-                    item.update(attribute)
-                    item.update({
-                        'task': 'copy'
-                    })
-                    items.append(item)
-                # Url
-                elif 'url' in attribute:
-                    item = {
-                        'unarchive': False
-                    }
-                    item.update(self._default(defaults, attribute['path']))
-                    item.update(attribute)
-                    item.update({
-                        'task': 'url'
-                    })
-                    items.append(item)
-                # File
-                else:
-                    item = {}
-                    item.update(self._default(defaults, attribute['path']))
-                    item.update(attribute)
-                    item.update({
-                        'task': 'file'
-                    })
-                    items.append(item)
+            # Must be a dict
+            if not isinstance(attribute, dict):
+                raise AnsibleError('Expect a dict but was a %s' % type(attribute))
 
-            # Merge by index key
+            # Check path key
+            if 'path' not in attribute:
+                raise AnsibleError('Expect a "path" key')
+
+            state = attribute.get('state')
+
+            # Ignore
+            if state == 'ignore':
+                continue
+
+            # Composite link
+            if state in ['link_directory', 'link_file']:
+                if 'src' not in attribute:
+                    raise AnsibleError('Expect "src" key')
+
+                # Composite
+                item = itemDefault.copy()
+                item.update(self._default(defaults, attribute['src']))
+                item.update(attribute)
+                item.update({
+                    'path': item.pop('src'),
+                    'state': state.split('_')[1]
+                })
+                items.append(item)
+
+                # Link
+                attribute['state'] = 'link'
+
+            item = itemDefault.copy()
+            item.update(self._default(defaults, attribute['path']))
+            item.update(attribute)
+            items.append(item)
+
+            # Merge by path key
             for item in items:
                 itemFound = False
                 for i, result in enumerate(results):